requirements.txt
Flask
openai==0.27.0
python-dotenv
anthropic
requests